backend/requirements.txt
langchain>=0.3
langchain-google-genai>=2.0
langgraph>=0.2
langsmith>=0.2
python-telegram-bot>=22.0
websockets>=13.0
aiohttp>=3.10.0
apscheduler>=3.10.0,<4.0.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
rich>=13.0
fastapi>=0.115
uvicorn>=0.30
